New posible nodes
Build road fully completed, as I know... As a first view, it seems correct, we didn't check every case, we will see in the future as we oould connect with the frontend.

In addition to add possible roads to build in the next turn for that player, now he can obtain the possible nodes to (with the correspondent condition to that.).